backend/pyproject.toml
[project]
name = "customer-service-agent"
version = "0.1.0"
description = "A customer service agent built with LiveKit, ElevenLabs, Deepgram, and Twilio."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "langfuse==2.60.8",
    "livekit-agents[deepgram,elevenlabs,openai,silero,turn-detector]",
    "python-dotenv",
    "uvicorn",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[dependency-groups]
dev = [
    "ruff>=0.12.0",
]
